requirements.txt
Python 
Torch
Transformers 
bitsandbytes
accelerate
Requests
Pillow